+++ src/app/layout.tsx
@@ -1,9 +1,14 @@
 import type { Metadata } from 'next';
-import { Rubik } from 'next/font/google';
+import { Cairo, Rubik } from 'next/font/google';
 import './globals.css';
 
 const rubik = Rubik({
   variable: '--font-rubik',
+  subsets: ['latin'],
+});
+
+const cairo = Cairo({
+  variable: '--font-cairo',
   subsets: ['latin'],
 });
 
@@ -19,7 +24,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="pt-BR">
-      <body className={`${rubik.variable} antialiased`}>{children}</body>
+      <body className={`${rubik.variable} ${cairo.variable} antialiased`}>
+        {children}
+      </body>
     </html>
   );
 }